Jet pt lower threshold and changing b-tagging sel.

+++ python/JetSelector_cfi.py
@@ -2,7 +2,7 @@ import FWCore.ParameterSet.Config as cms
 
 defaultJetSelectionParameters = cms.PSet(
     jettype             = cms.string('AK5JET'),
-    jetPtMin            = cms.double(50),
+    jetPtMin            = cms.double(20),
     jetPtMax            = cms.double(100000),
     jetAbsEtaMax        = cms.double(2.4),
     jetCSVDiscMin       = cms.double(-10000.),
